requirements.txt
fastapi
uvicorn
httpx
requests
pydantic
python-multipart
pdfplumber
PyMuPDF
sentence-transformers
faiss-cpu
huggingface-hub
python-dotenv
google-genai
rank-bm25
regex
python-docx
langchain
pinecone
pinecone-text